% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.endid}
\alias{print.endid}
\title{Print an endid object}
\usage{
\method{print}{endid}(x, ...)
}
\arguments{
\item{x}{An \code{endid} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an endid object
}
