% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.endid}
\alias{print.summary.endid}
\title{Print a summary.endid object}
\usage{
\method{print}{summary.endid}(x, ...)
}
\arguments{
\item{x}{A \code{summary.endid} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a summary.endid object
}
