% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.endid}
\alias{summary.endid}
\title{Summarize an endid object}
\usage{
\method{summary}{endid}(object, ...)
}
\arguments{
\item{object}{An \code{endid} object.}

\item{...}{Ignored.}
}
\value{
A \code{summary.endid} object (list with att_table and qte_table).
}
\description{
Summarize an endid object
}
